requirements.txt
backcall==0.2.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
future==0.18.2
ipykernel==5.3.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==6.1.6
jupyter-core==4.6.3
kiwisolver==1.2.0
matplotlib==3.3.0
numpy==1.19.0
opencv-python==4.3.0.36
parso==0.7.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pyzmq==19.0.1
scipy==1.5.1
six==1.15.0
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.2.5
